pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-wiki"
version = "0.1.0"
description = "Game-Aware RAG Agent based on LangChain and MCP"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "langchain-openai>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    "defusedxml>=0.7.1",
    "qdrant-client>=1.7.0",
    "rank_bm25>=0.2.2",
    "python-dotenv>=1.0.1",
    "langchain>=1.2.12",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-classic>=1.0.3",
    "langchain-google-genai>=1.0.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "langgraph-cli"
]

[tool.setuptools.packages.find]
where = ["src"]
